refactor(test): migrate org.ts helpers to three-tier test architecture (#9373)
Move 17 DB-direct functions from api-test-helpers/org.ts into
db-test-seeders/org.ts (11 seeders) and db-test-assertions/org.ts
(6 assertions), then refactor org.ts to pure re-exports.
